+++ NEWS.md
@@ -1,6 +1,7 @@
 # states 0.2.2.9000
 
-- Add country_names() to translate country codes to country names. 
+- Add country_names() to translate country codes to country names and pretty_names() to shorten some of the longer country names, like "Macedonia, the former Yugoslav Republic of".
+- Updated vignette on G&W and COW differences to take into account when trying to translate between codes. 
 
 # states 0.2.2